chore(deps): break utils/colors cycles via leaf types module
Extracts CategoryColorDefinition + EQUIPMENT_CATEGORY_COLORS into a
new leaf module equipmentColors.types.ts. Both the ammo sub-type
detector and the legacy compatibility layer now import the shared
type and category map from the leaf rather than from the orchestrator,
which was previously importing them back via re-export.

Cycles closed:
  1) equipmentColors.ammo.ts -> equipmentColors.ts
  2) equipmentColors.ts -> equipmentColors.legacy.ts

equipmentColors.ts continues to re-export the type and map for
backward compatibility, so all existing consumers are unaffected.

No behavior change. Public API surface (exports from
@/utils/colors/equipmentColors) is identical.
